+++ lib/active_record/connection_adapters/nulldb_adapter/checkpoint.rb
@@ -1,0 +1,13 @@
+class ActiveRecord::ConnectionAdapters::NullDBAdapter
+
+  class Checkpoint < Statement
+    def initialize
+      super(:checkpoint, "")
+    end
+
+    def ==(other)
+      self.class == other.class
+    end
+  end
+
+end
