+++ lib/active_record/connection_adapters/nulldb_adapter/configuration.rb
@@ -1,0 +1,5 @@
+class ActiveRecord::ConnectionAdapters::NullDBAdapter
+
+  class Configuration < Struct.new(:project_root); end
+
+end
