+++ lib/active_record/connection_adapters/nulldb_adapter/core.rb
@@ -1,118 +1,4 @@
-require 'logger'
-require 'stringio'
-require 'singleton'
-require 'pathname'
-require 'active_support'
-require 'active_record/connection_adapters/abstract_adapter'
-require 'nulldb/core'
-
-unless respond_to?(:tap)
-  class Object
-    def tap
-      yield self
-      self
-    end
-  end
-end
-
-unless respond_to?(:try)
-  class Object
-    def try(*a, &b)
-      if a.empty? && block_given?
-        yield self
-      else
-        __send__(*a, &b)
-      end
-    end
-  end
-
-  class NilClass
-    def try(*args); nil; end
-  end
-end
-
-class ActiveRecord::Base
-  # Instantiate a new NullDB connection.  Used by ActiveRecord internally.
-  def self.nulldb_connection(config)
-    ActiveRecord::ConnectionAdapters::NullDBAdapter.new(config)
-  end
-end
-
-
-module ActiveRecord
-  # Just make sure you have the latest version of your schema
-  class Schema < Migration
-    def self.define(info={}, &block)
-      instance_eval(&block)
-    end
-  end
-end
-
-
-class ActiveRecord::ConnectionAdapters::NullDBAdapter <
-    ActiveRecord::ConnectionAdapters::AbstractAdapter
-
-  class Column < ::ActiveRecord::ConnectionAdapters::Column
-    private
-
-    def simplified_type(field_type)
-      type = super
-      type = :integer if type.nil? && sql_type == :primary_key
-      type
-    end
-  end
-
-  class Statement
-    attr_reader :entry_point, :content
-
-    def initialize(entry_point, content = "")
-      @entry_point, @content = entry_point, content
-    end
-
-    def ==(other)
-      self.entry_point == other.entry_point
-    end
-  end
-
-  class Checkpoint < Statement
-    def initialize
-      super(:checkpoint, "")
-    end
-
-    def ==(other)
-      self.class == other.class
-    end
-  end
-
-  TableDefinition = ActiveRecord::ConnectionAdapters::TableDefinition
-
-  class IndexDefinition < Struct.new(:table, :name, :unique, :columns, :lengths, :orders)
-  end
-
-  class NullObject
-    def method_missing(*args, &block)
-      nil
-    end
-
-    def to_a
-      []
-    end
-  end
-
-  class EmptyResult < Array
-    attr_writer :columns
-    def rows
-      []
-    end
-
-    def column_types
-      columns.map{|col| col.type}
-    end
-
-    def columns
-      @columns ||= []
-    end
-  end
+class ActiveRecord::ConnectionAdapters::NullDBAdapter < ActiveRecord::ConnectionAdapters::AbstractAdapter
 
   # A convenience method for integratinginto RSpec.  See README for example of
   # use.
