+++ lib/active_record/connection_adapters/nulldb_adapter/empty_result.rb
@@ -1,0 +1,18 @@
+class ActiveRecord::ConnectionAdapters::NullDBAdapter
+
+  class EmptyResult < Array
+    attr_writer :columns
+    def rows
+      []
+    end
+
+    def column_types
+      columns.map{|col| col.type}
+    end
+
+    def columns
+      @columns ||= []
+    end
+  end
+
+end
