+++ lib/active_record/connection_adapters/nulldb_adapter/index_definition.rb
@@ -1,0 +1,5 @@
+class ActiveRecord::ConnectionAdapters::NullDBAdapter
+
+  class IndexDefinition < Struct.new(:table, :name, :unique, :columns, :lengths, :orders); end
+
+end
