+++ lib/active_record/connection_adapters/nulldb_adapter/null_object.rb
@@ -1,0 +1,13 @@
+class ActiveRecord::ConnectionAdapters::NullDBAdapter
+
+  class NullObject
+    def method_missing(*args, &block)
+      nil
+    end
+
+    def to_a
+      []
+    end
+  end
+
+end
