+++ lib/active_record/connection_adapters/nulldb_adapter/table_definition.rb
@@ -1,0 +1,5 @@
+class ActiveRecord::ConnectionAdapters::NullDBAdapter
+
+  TableDefinition = ActiveRecord::ConnectionAdapters::TableDefinition
+
+end
